--- a/test/performance/sketch/CMakeLists.txt
+++ b/test/performance/sketch/CMakeLists.txt
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2006-2023, Knut Reinert & Freie Universität Berlin
+# SPDX-FileCopyrightText: 2016-2023, Knut Reinert & MPI für molekulare Genetik
+# SPDX-License-Identifier: BSD-3-Clause
+
+hibf_benchmark (hyperloglog_benchmark.cpp)