Iterations are in thousands

| Benchmark          | Bits | Iterations before | Iterations after | Speedup |
|--------------------|------|-------------------|------------------|---------|
| add                | 5    | 303,112           | 302,424          | 1.0     |
| add                | 6    | 312,195           | 301,988          | 1.0     |
| add                | 7    | 307,077           | 304,696          | 1.0     |
| add                | 8    | 311,200           | 311,102          | 1.0     |
| add                | 9    | 311,268           | 310,932          | 1.0     |
| add                | 10   | 311,394           | 310,958          | 1.0     |
| add                | 11   | 311,388           | 311,060          | 1.0     |
| add                | 12   | 311,417           | 311,068          | 1.0     |
| merge              | 5    | 1,000,000         | 1,000,000        | 1.0     |
| merge              | 6    | 1,000,000         | 1,000,000        | 1.0     |
| merge              | 7    | 814,029           | 813,645          | 1.0     |
| merge              | 8    | 215,687           | 213,490          | 1.0     |
| merge              | 9    | 220,560           | 221,890          | 1.0     |
| merge              | 10   | 128,382           | 128,417          | 1.0     |
| merge              | 11   | 69,709            | 69,652           | 1.0     |
| merge              | 12   | 36,424            | 36,395           | 1.0     |
| estimate           | 5    | 20,056            | 37,375           | 1.9     |
| estimate           | 6    | 7,558             | 21,712           | 2.9     |
| estimate           | 7    | 3,703             | 12,470           | 3.4     |
| estimate           | 8    | 1,737             | 6,894            | 4.0     |
| estimate           | 9    | 869               | 3,576            | 4.1     |
| estimate           | 10   | 435               | 1,824            | 4.2     |
| estimate           | 11   | 217               | 916              | 4.2     |
| estimate           | 12   | 109               | 462              | 4.2     |
| merge_and_estimate | 5    | 19,685            | 36,999           | 1.9     |
| merge_and_estimate | 6    | 7,460             | 21,518           | 2.9     |
| merge_and_estimate | 7    | 3,665             | 12,617           | 3.4     |
| merge_and_estimate | 8    | 1,731             | 6,856            | 4.0     |
| merge_and_estimate | 9    | 865               | 3,565            | 4.1     |
| merge_and_estimate | 10   | 433               | 1,818            | 4.2     |
| merge_and_estimate | 11   | 217               | 913              | 4.2     |
| merge_and_estimate | 12   | 108               | 458              | 4.2     |